requirements.txt
torch
torchvision
numpy
pillow
matplotlib
tqdm
scikit-image
pyyaml
einops
timm
